First commit... compilation ok, but crash after lunch (
