+++ MovieQuiz/Helpers/UIColor+Extensions.swift
@@ -2,3 +2,10 @@ import Foundation
 import UIKit
 
 extension UIColor { }
+
+extension UIColor{
+    static var ypGreen: UIColor {UIColor (named: "YP Green")!}
+    static var ypRed: UIColor {UIColor (named: "YP Red")!}
+    
+    
+}
