+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,16 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Alexander Farizanov on 02.12.2022.
+//
+
+import Foundation
+import UIKit
+
+struct QuizQuestion {
+    let image: String
+    //let image: UIImage
+    let text: String
+    let correctAnswer: Bool
+}
