+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Alexander Farizanov on 02.12.2022.
+//
+
+import Foundation
+import UIKit
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
