+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Alexander Farizanov on 02.12.2022.
+//
+
+import Foundation
+import UIKit
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+    
+}
